requirements.txt
django==1.7
PyYAML==3.11
djangorestframework==2.4.3
